Add ability to notify app_store
