+++ app/presenters/check_answers/work_items_card.rb
@@ -25,14 +25,6 @@ module CheckAnswers
       header_rows + work_item_rows + total_rows
     end
 
-    def as_json(*)
-      claim.work_items.map do |work_item|
-        data = work_item.attributes.slice!('claim_id', 'created_at', 'updated_at')
-        data['completed_on'] = data['completed_on'].to_s
-        data
-      end
-    end
-
     private
 
     def header_rows
